feat(*): update utils exports and rollup config
