+++ packages/hooks/package.json
@@ -11,7 +11,7 @@
     ],
     "scripts": {
         "update:exports": "node ../../scripts/update-exports.js",
-        "build": "yarn build:ts && yarn build:dts",
+        "build": "yarn build:ts && yarn build:dts && yarn update:exports",
         "build:ts": "rollup -c",
         "build:dts": "dts-bundle-generator --no-banner --no-check --config dts.json"
     },
