+++ packages/utils/src/crop-account-number/index.ts
@@ -1,1 +1,10 @@
-export * from './util';
+/**
+ * Возвращает 4 последние цифры номера счёта в формате `··XXXX`
+ */
+export function cropAccountNumber(accountNumber: string) {
+    if (!accountNumber) {
+        return accountNumber;
+    }
+
+    return `\u00B7\u00B7${accountNumber.slice(-4)}`;
+}
