+++ packages/utils/src/get-all-currency-codes/index.ts
@@ -1,1 +1,5 @@
-export * from './util';
+import { currency } from '@alfalab/data';
+
+export function getAllCurrencyCodes() {
+    return Object.keys(currency.CURRENCY_SYMBOLS);
+}
