--- packages/utils/src/get-all-currency-codes/util.ts
@@ -1,5 +1,0 @@
-import { currency } from '@alfalab/data';
-
-export function getAllCurrencyCodes() {
-    return Object.keys(currency.CURRENCY_SYMBOLS);
-}
