+++ packages/utils/src/get-currency-symbol/index.ts
@@ -1,1 +1,9 @@
-export * from './util';
+import { currency, CurrencyCodes } from '@alfalab/data';
+
+/**
+ * Возвращает знак валюты по ISO коду.
+ *
+ * @param currencyCode Код валюты.
+ */
+export const getCurrencySymbol = (currencyCode: CurrencyCodes): string =>
+    currency.CURRENCY_SYMBOLS[currencyCode];
