--- packages/utils/src/get-currency-symbol/util.ts
@@ -1,9 +1,0 @@
-import { currency, CurrencyCodes } from '@alfalab/data';
-
-/**
- * Возвращает знак валюты по ISO коду.
- *
- * @param currencyCode Код валюты.
- */
-export const getCurrencySymbol = (currencyCode: CurrencyCodes): string =>
-    currency.CURRENCY_SYMBOLS[currencyCode];
