+++ packages/utils/src/has-scrolled-to-bottom-of-page/index.ts
@@ -1,1 +1,10 @@
-export * from './util';
+/**
+ * Возвращает TRUE, если проскролено до низа страницы
+ */
+export function hasScrolledToBottomOfPage() {
+    const { documentElement } = document;
+    const offset = documentElement.scrollTop + window.innerHeight;
+    const height = documentElement.offsetHeight;
+
+    return offset >= height;
+}
