+++ packages/utils/src/is-overflown/index.test.ts
@@ -1,4 +1,4 @@
-import { isOverflown } from './';
+import { isOverflown } from '.';
 
 describe('isOverflown', () => {
     it('should return `true` if element content height overflowing', () => {
