+++ packages/utils/src/is-overflown/index.ts
@@ -1,1 +1,7 @@
-export * from './util';
+/**
+ * Возвращает `true`, если элемент переполнен
+ * @param element HTML-элемент
+ */
+export function isOverflown(element: HTMLElement) {
+    return element.scrollHeight > element.clientHeight || element.scrollWidth > element.clientWidth;
+}
