--- packages/utils/src/is-overflown/util.ts
@@ -1,7 +1,0 @@
-/**
- * Возвращает `true`, если элемент переполнен
- * @param element HTML-элемент
- */
-export function isOverflown(element: HTMLElement) {
-    return element.scrollHeight > element.clientHeight || element.scrollWidth > element.clientWidth;
-}
