+++ packages/utils/src/pluralize/index.test.ts
