+++ packages/utils/src/pluralize/index.ts
@@ -1,1 +1,26 @@
-export * from './util';
+export function pluralize(
+  number: number,
+  one: string,
+  two: string,
+  five: string
+): string {
+  let n = Math.abs(number);
+
+  n %= 100;
+
+  if (n >= 5 && n <= 20) {
+    return five;
+  }
+
+  n %= 10;
+
+  if (n === 1) {
+    return one;
+  }
+
+  if (n >= 2 && n <= 4) {
+    return two;
+  }
+
+  return five;
+}
